internal: remove cast to int in NGLI_ARRAY_NB

+++ libnopegl/src/android_ctx.c
@@ -113,7 +113,7 @@ static int has_native_imagereader_api_support(struct gpu_ctx *gpu_ctx)
             VK_EXT_QUEUE_FAMILY_FOREIGN_EXTENSION_NAME,
             VK_ANDROID_EXTERNAL_MEMORY_ANDROID_HARDWARE_BUFFER_EXTENSION_NAME,
         };
-        for (int i = 0; i < NGLI_ARRAY_NB(required_extensions); i++) {
+        for (size_t i = 0; i < NGLI_ARRAY_NB(required_extensions); i++) {
             if (!ngli_vkcontext_has_extension(vk, required_extensions[i], 1))
                 return 0;
         }
